Assigning BDFL delegate for JEP-202
For the context of how I select BDFL delegate, please see here:
https://groups.google.com/forum/#!msg/jenkinsci-dev/spDAr8EJm3c/ba-ppaM7BgAJ

This part of the core has been worked on by many people, and any of them
can be a BDFL delegate to this JEP. That said, Jesse has worked more
recently on the artifact manager API, and I know he's involved in this
effort quite extensively. So I think he's the best person to ensure
consensus on this effort.

Signed-off-by: Kohsuke Kawaguchi <kk@kohsuke.org>