Accepting JEP #6 as BDFL.
In retrospect, I feel like this JEP is squarely in the authority of the
security officer, and I should have just assigned Daniel as BDFL
Delegate. I apologize for any time delay not doing that might have
caused. I expect the future JEPs in this space to have the security
officer at that time as the initial BDFL Delegate.

The discussion I saw in the mailing list was all supportive of this
move, and my hats off for Daniel's continued push to better security
handling in the Jenkins project.

Signed-off-by: Kohsuke Kawaguchi <kk@kohsuke.org>