Assigning Oleg as BDFL Delegatej
This is more of a process within GSoC, which is run by Oleg. From that
perspective, this buck needs to stop at him. As a BDFL delegate, I
expect him to make sure this is in line with the existing budget/expense
process by talking to the board -- I will send him a separate note on
that.

Signed-off-by: Kohsuke Kawaguchi <kk@kohsuke.org>