Update the title of Jep 229 to include plugins.
The title of JEP-229 confused me as it was talking about Jenkins
Components yet is equally applicaple to Jenkins plugins.

Normally a component would be something that goes into Jenkins like
`remoting` and one would not typically refer to a Plugin as a component.

This updates the title and the JEP body in a few places to make it clear
this is not just for components.