Adjust transition period for Java 17 as minimum version
The original Google Doc and the implementation of the administrative
monitor both describe an approximtely 18 month period where Java 17 is
the minimum Java version required for Jenkins.

The first commits to this JEP described an approximately 12 month period
where Java 17 is the minimum Java version required for Jenkins.

This change corrects the JEP to use an approximately 18 month period
where Java 17 is the minimum Java version required for Jenkins.  In order
to do that, Java 21 will be the the minimum Java version required for
Jenkins for an 18 month period as well.

Old durations of minimum Java versions in this document were:

* Java 11 - 25 months
* Java 17 - 12 months
* Java 21 - 24 months
* Java 25 - 24 months

New durations of minimum Java versions in this document are:

* Java 11 - 25 months
* Java 17 - 18 months
* Java 21 - 18 months
* Java 25 - 24 months

The new duration provides a more gradual transition to the 2+2+2 Java
support plan.

This commit also adds a description of the Java 25 (and beyond) steady
state.